inicio do grafico em c_s

+++ src/com/fortes/rh/model/relatorio/DataGrafico.java
@@ -3,9 +3,9 @@ package com.fortes.rh.model.relatorio;
 public class DataGrafico {
 	
 	private String label;
-	private int data;
+	private double data;
 	
-	public DataGrafico(String label, int data) {
+	public DataGrafico(String label, double data) {
 		super();
 		this.label = label;
 		this.data = data;
@@ -17,10 +17,12 @@ public class DataGrafico {
 	public void setLabel(String label) {
 		this.label = label;
 	}
-	public int getData() {
+
+	public double getData() {
 		return data;
 	}
-	public void setData(int data) {
+
+	public void setData(double data) {
 		this.data = data;
 	}
